Add files via upload
This executable file is designed to streamline the process of downloading Sysmon files, installing the Sysmon service, and configuring the ossec.conf file for your installed Wazuh agent. Simply download and run the file, and everything will be set up automatically.

The downloaded files will be stored on your desktop in the "Sysmon Files" folder.